requirements.txt
aiogram==3.13.1
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.2
python-dotenv==1.0.1
